requirements.txt
asgiref==3.7.2
astroid==2.15.5
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.8
dill==0.3.6
Django==4.2.1
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djlint==1.30.2
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
isort==5.12.0
jsbeautifier==1.14.8
json5==0.9.14
lazy-object-proxy==1.9.0
mccabe==0.7.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
tqdm==4.65.0
wrapt==1.15.0
